refactor(auth): update delete user(account)

+++ next-app/src/core/auth/actions/delete-user.ts
@@ -1,21 +1,28 @@
 "use server";
 
+import { MESSAGES } from "@/constants/messages";
 import { auth } from "@/lib/auth";
 import { APIError } from "better-auth";
 import { revalidatePath } from "next/cache";
 import { headers } from "next/headers";
+import z from "zod";
+import { DeleteAccountSchema } from "../schemas/delete-account";
 
-type DeleteUserResponse =
-  | {
-      success: string;
-      error?: null;
-    }
-  | {
-      success?: null;
-      error: string;
+export const deleteUser = async (
+  values: z.infer<typeof DeleteAccountSchema>,
+  userEmail: string,
+) => {
+  const validatedFields = DeleteAccountSchema.safeParse(values);
+
+  if (!validatedFields.success) return { error: MESSAGES.INVALID_FIELDS };
+
+  const { email } = validatedFields.data;
+
+  if (userEmail !== email)
+    return {
+      error: MESSAGES.EMAIL_OLD_NOT_MATCH,
     };
 
-export const deleteUser = async (): Promise<DeleteUserResponse> => {
   try {
     const data = await auth.api.deleteUser({
       body: {
